forecasting-example/components/data_ingestion/requirements.txt
google-cloud-storage
fire